requirements.txt
flask
requests
python-dotenv
pydantic
openai
gunicorn
